+++ python/taichi/type/primitive_types.py
@@ -6,13 +6,13 @@ float32 = ti_core.DataType_f32
 """32-bit single precision floating point data type.
 """
 f32 = float32
-"""Alias for :const:`~taichi.types.primitive_types.float32`
+"""Alias for :const:`~taichi.type.primitive_types.float32`
 """
 float64 = ti_core.DataType_f64
 """64-bit double precision floating point data type.
 """
 f64 = float64
-"""Alias for :const:`~taichi.types.primitive_types.float64`
+"""Alias for :const:`~taichi.type.primitive_types.float64`
 """
 
 real_types = [f32, f64, float]
@@ -28,13 +28,13 @@ int32 = ti_core.DataType_i32
 """32-bit signed integer data type.
 """
 i32 = int32
-"""Alias for :const:`~taichi.types.primitive_types.int32`
+"""Alias for :const:`~taichi.type.primitive_types.int32`
 """
 int64 = ti_core.DataType_i64
 """64-bit signed integer data type.
 """
 i64 = int64
-"""Alias for :const:`~taichi.types.primitive_types.int64`
+"""Alias for :const:`~taichi.type.primitive_types.int64`
 """
 
 uint8 = ti_core.DataType_u8
@@ -45,13 +45,13 @@ uint32 = ti_core.DataType_u32
 """32-bit unsigned integer data type.
 """
 u32 = uint32
-"""Alias for :const:`~taichi.types.primitive_types.uint32`
+"""Alias for :const:`~taichi.type.primitive_types.uint32`
 """
 uint64 = ti_core.DataType_u64
 """64-bit unsigned integer data type.
 """
 u64 = uint64
-"""Alias for :const:`~taichi.types.primitive_types.uint64`
+"""Alias for :const:`~taichi.type.primitive_types.uint64`
 """
 
 integer_types = [i8, i16, i32, i64, u8, u16, u32, u64, int]
