+++ python/taichi/ui/utils.py
@@ -7,7 +7,7 @@ from taichi.lang.kernel_arguments import ext_arr, template
 from taichi.lang.kernel_impl import kernel
 from taichi.lang.matrix import Vector
 from taichi.lang.ops import get_addr
-from taichi.types.primitive_types import u64
+from taichi.type.primitive_types import u64
 
 
 @kernel
